fix(build): fixes compatibility of module exports and npm
Updates scripts to be compatible with non-yarn setups. Updates exports to also export modules
individually.
