+++ package.json
@@ -3,13 +3,13 @@
   "version": "2.1.3",
   "description": "🔖 Generates 1D, 2D, and composite barcodes in png, svg, or eps formats. Supports 50+ symbologies.",
   "scripts": {
-    "commit": "yarn git-cz",
+    "commit": "git-cz",
     "release": "standard-version",
-    "lint": "yarn eslint ./src",
+    "lint": "eslint ./src",
     "test:e2e": "rimraf test/e2e/__rendered__/* && jest test/e2e",
     "test:unit": "jest src --collectCoverage true",
-    "test": "yarn test:unit && yarn test:e2e",
-    "build": "yarn tsc && npm run binary:rebuild",
+    "test": "npm run test:unit && npm run test:e2e",
+    "build": "tsc && npm run binary:rebuild",
     "binary:install": "node-pre-gyp install --fallback-to-build",
     "binary:rebuild": "node-pre-gyp rebuild package",
     "install": "node ./scripts/install.js && npm run binary:install",
@@ -79,7 +79,7 @@
     "vuepress": "^1.8.2"
   },
   "bundledDependencies": [
-    "node-pre-gyp"
+    "@mapbox/node-pre-gyp"
   ],
   "husky": {
     "hooks": {
